requirements.txt
dash-labs==0.2.0
plotly_express==0.4.0
koalas==1.8.1
openpyxl==3.0.7
pandas==1.3.1
pyspark==3.0.1
urllib3==1.26.6
websocket-client==1.1.0
py4j==0.10.9
pyarrow==5.0.0
